задание_повышенной_сложности_3/python_client/requirements.txt
PyJWT>=2.8.0
requests>=2.31.0
responses>=0.24.0
pytest>=7.4.3
